move SymplifyCookie code to correct namespace
